refactor: rename to persist.ts
